Move `TranslationDynamicScope` out of `expressions` package
